requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
joblib
python-multipart
firebase-admin

